requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
dnspython==2.2.0
fastapi==0.74.1
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
h11==0.13.0
httptools==0.3.0
idna==3.3
impacket==0.9.24
itsdangerous==2.1.0
Jinja2==3.0.3
ldap3==2.9.1
ldapdomaindump==0.9.3
MarkupSafe==2.1.0
maxminddb==2.2.0
multidict==6.0.2
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.14.1
pydantic==1.9.0
pyOpenSSL==22.0.0
pypsexec==0.3.0
pyspnego==0.5.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
smbprotocol==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
urlopen==1.0.0
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
Werkzeug==2.0.3
yarl==1.7.2
